# internal/api/user-auth.go
package api

import (
	"encoding/json"
	"fmt"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/sirupsen/logrus"
	"goph_keeper/pkg/auth"
	"goph_keeper/pkg/common"
	"io"
	"net/http"
	"strings"
)

type UserAuthRequest struct {
	Login    string `json:"login"`
	Password string `json:"password"`
}

// UserAuth godoc
// @Summary      Авторизация пользователя
// @Description  Авторизация пользователя
// @Tags         register
// @Accept       json
// @Produce      json
// @Param        request body UserAuthRequest true "Параметры запроса"
// @Failure      500
// @Router       /api/v1/user/login [post]
func (h *Handler) UserAuth(w http.ResponseWriter, r *http.Request) {
	logger := logrus.WithFields(map[string]interface{}{
		"HANDLER":    "UserAuth",
		"REQUEST_ID": r.Context().Value(middleware.RequestIDKey),
	})
	var userAuthRequest UserAuthRequest
	bodyBytes, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte(fmt.Sprintf("failed to read request body. err: %v", err)))
		return
	}
	err = json.Unmarshal(bodyBytes, &userAuthRequest)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte(fmt.Sprintf("failed to read request body. err: %v", err)))
		return
	}
	if !validateUserAuthRequest(userAuthRequest) {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte("validation failed"))
		return
	}

	user, err := h.Storage.GetUserByLogin(r.Context(), userAuthRequest.Login)
	if err != nil || user == nil {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte(fmt.Sprintf("user not found. err: %v", err)))
		return
	}
	if !h.Auth.CompareHashAndPassword(userAuthRequest.Password, user.Password) {
		logger.Errorf("compare hash and password fail for user: %s", userAuthRequest.Login)
		w.WriteHeader(http.StatusUnauthorized)
		return
	}
	token, err := h.Auth.CreateToken(&auth.User{
		ID:       user.ID,
		Login:    user.Login,
		Password: user.Password,
		Phrase:   user.Phrase,
	})
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		w.Write([]byte(fmt.Sprintf("could not create token. err: %v", err)))
		return
	}
	logger.Debugf("user login:`%s` password:`%s` token:`%s` created", user.Login, user.Password, token)
	w.Header().Set(common.AuthorizationHeaderName, fmt.Sprintf("Bearer %s", string(token)))

}

func validateUserAuthRequest(request UserAuthRequest) bool {
	if strings.TrimSpace(request.Login) != "" && strings.TrimSpace(request.Password) != "" {
		return true
	}
	return false
}