# internal/api/user-register.go
package api

import (
	"encoding/json"
	"fmt"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/sirupsen/logrus"
	"goph_keeper/pkg/auth"
	"goph_keeper/pkg/common"
	"io"
	"net/http"
)

type UserRegisterRequest struct {
	Login    string `json:"login"`
	Password string `json:"password"`
	Phrase   string `json:"phrase"`
}

// UserRegister godoc
// @Summary      Зарегистрировать пользователя
// @Description  Зарегистрировать пользователя
// @Tags         register
// @Accept       json
// @Produce      json
// @Param        request body UserRegisterRequest true "Параметры запроса"
// @Failure      500
// @Router       /api/v1/user/register [post]
func (h *Handler) UserRegister(w http.ResponseWriter, r *http.Request) {
	logger := logrus.WithFields(map[string]interface{}{
		"HANDLER":    "UserRegister",
		"REQUEST_ID": r.Context().Value(middleware.RequestIDKey),
	})
	var userRegisterRequest UserRegisterRequest
	bodyBytes, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte(fmt.Sprintf("failed to read request body. err: %v", err)))
		return
	}
	err = json.Unmarshal(bodyBytes, &userRegisterRequest)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest)
		w.Write([]byte(fmt.Sprintf("failed to read request body. err: %v", err)))
		return
	}
	hashPassword, err := h.Auth.CreateHash(userRegisterRequest.Password)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		w.Write([]byte("could not create hash for password"))
		return
	}
	user, err := h.Storage.CreateUser(r.Context(), userRegisterRequest.Login, hashPassword, userRegisterRequest.Phrase)
	if err != nil {
		w.WriteHeader(http.StatusConflict)
		logger.Errorf("user login:`%s` not saved. err:`%s`", userRegisterRequest.Login, err)
		w.Write([]byte(fmt.Sprintf("user login:`%s` not saved", userRegisterRequest.Login)))
		return
	}
	token, err := h.Auth.CreateToken(&auth.User{
		ID:       user.ID,
		Login:    user.Login,
		Password: user.Password,
		Phrase:   user.Phrase,
	})
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		w.Write([]byte(fmt.Sprintf("could not create token. err: %v", err)))
		return
	}
	logger.Debugf("user login:`%s` password:`%s` token:`%s` created", user.Login, user.Password, token)
	w.Header().Set(common.AuthorizationHeaderName, fmt.Sprintf("Bearer %s", string(token)))
}